Fix metric computation with Metric objects in AdversarialRegularization.
Cause of the problem:

When a Keras model calls `self.add_metric()` with `aggregation='mean'`, a new
`Metric` object is created to track the state of the metric. But if the metric
is calculated by an existing `Metric` object, the state update only happens on
the existing `Metric` object, but not the newly created one. Thus the shown
metric (from the new `Metric` object) is never updated.

Details of the fix:

The fix sets `aggregation` to `None` when a metric is calculated by a `Metric`
object.  In this way the Keras model will report the numbers from that `Metric`
object, instead of creating a new one. But another issue is triggered: The same
`Metric` object is now compiled into both the base model and the
adversarial-regularized model, which messes up the computational graph when
running with Tensorflow 1.x (graph mode).

To resolve the computational graph issue, each `Metric` object is cloned before
passing to the base model. The `Metric` objects are also cloned if they are
to be applied on multiple outputs.

Fixes #8

PiperOrigin-RevId: 269422315