Unset Loss.reduction to prevent double-reduction in AdversarialRegularization.
`AdversarialRegularization` creates a loss wrapper around the provided loss in
`compile()` for handling sample weights and loss reduction (aggregation). If the
provided loss is a `tf.keras.losses.Loss` object, it comes with loss reduction
by default which causes an error in the loss wrapper because the wrapper expects
unreduced loss values.

This change disables the loss reduction in the provided `Loss` object, so the
loss wrapper can function properly. An alternative approach would be disabling
the loss reduction in the loss wrapper while doing the loss reduction in the
`Loss` object. However, the alternative approach would run into an error when
running with `tf.distribute.Strategy`, because the `SUM_OVER_BATCH_SIZE`
reduction type requires special logic outside the `Loss` object. Such logic is
already implemented in the loss wrapper, so letting the wrapper handle loss
reduction looks cleaner.

Fixes #21

PiperOrigin-RevId: 272923076