Changes the graph builder to use far less memory.
In particular, it will no longer store the entire result
graph in memory, so the memory consumption is proportional
only to the size of the input, not the size of the input
plus the size of the output.

Because graph building and writing are now combined, the
log messages produced by the build_graph function/binary
have changed slightly. The program also now produces a
log message every 1M edges it writes to indicate progress.

PiperOrigin-RevId: 276354066