Adds a new --lsh_bits flag to nsl.tools.build_graph().
If --lsh_bits > 0, all the input points are first bucketed into
at most 2^n buckets, where 'n' is the --lsh_bits value. Each of
the buckets is then considered independently, vastly reducing
the number of pairs of points that need to be compared for
similarity. Thus, the running time of the program can be reduced
dramatically when the number of input points is large, at the
potential expense of missing some graph edges.

This also introduces a new configuration class and a new top-level function:
  nsl.configs.GraphBuilderConfig: graph building configuration object.
  nsl.tools.build_graph_from_config(...): parameterized by GraphBuilderConfig.

PiperOrigin-RevId: 318376805