Graph regularization for Twitter rumour veracity classification using natural graphs
Overview
This tutorial uses the PHEME dataset for veracity classification of Twitter rumours, a binary classification task.

Tweet texts are used as input for computing embeddings, e.g. with ALBERT. Those representations are then used as features for the baseline MLP classification model, as well as for the graph regularized version, which uses the structure defined by the tweet replies as a natural graph.

Attribution
We use the PHEME dataset for Rumour Detection and Veracity Classification created by Elena Kochkina, Maria Liakata, and Arkaitz Zubiaga.

This data set is licensed under CC BY 4.0. The underlying data is used without modifications as labels or for computing model features.

PiperOrigin-RevId: 323743221