Support Sequential models with DenseFeatures in AdversarialRegularization.
The issue (raised in #74) is that if the first layer of a `Sequential` model is
a `DenseFeatures` layer, `model.input_names` is set to `None`. But
`AdversarialRegularization` relies on the input names to indicate the expected
features of the base model.

With this change, `AdversarialRegularization` will not try to flatten the
features when the base model is `Sequential` and its `input_names` is `None`.

PiperOrigin-RevId: 336185318