Fix heap misusage for top-k, top-n.
The original code relied on an implementation detail and ignored the precondition that [first, last) be in heap order. The std::push_heap call was assumed to act solely as a front/back-swap and sift-down operation. This now fails in an upcoming version of clang.

The standard-adhering operation is
```
std::pop_heap(heap.begin(), heap.end(), comp);
heap.back() = v;
std::push_heap(heap.begin(), heap.end(), comp);
```
Also adjusted container bounds to no longer include an unused element in the back, since it is no longer required and complicates the above replacement.
PiperOrigin-RevId: 436894332